Compiler: Pack colors using UINT8 rather than UINT16
This is because the highest the color components will ever go in value
is 100.

+++ compiler/components/DNACornice.py
@@ -24,5 +24,5 @@ class DNACornice(DNAGroup.DNAGroup):
         packer.pack('code', self.code, SHORT_STRING)
 
         for component in self.color:
-            packer.pack('color', int(component * 255), UINT16)
+            packer.pack('color', int(component * 255), UINT8)
         return packer
